Create Django-React App

+++ src/containers/ArticleDetailView.js
@@ -1,6 +1,6 @@
 import React from "react";
 import axios from "axios";
-
+import { connect } from "react-redux";
 import { Button, Card, Form } from "antd";
 import CustomForm from "../components/Form";
 
@@ -9,20 +9,34 @@ class ArticleDetail extends React.Component {
     article: {}
   };
 
-  componentDidMount() {
-    const articleID = this.props.match.params.articleID;
-    axios.get(`http://127.0.0.1:8000/api/${articleID}/`).then(res => {
-      this.setState({
-        article: res.data
+  componentWillReceiveProps(newProps) {
+    if (newProps.token) {
+      axios.defaults.headers = {
+        "Content-Type": "application/json",
+        Authorization: newProps.token
+      };
+      axios.get("http://127.0.0.1:8000/api/").then(res => {
+        this.setState({
+          articles: res.data
+        });
+        console.log(res.data);
       });
-    });
+    }
   }
 
   handleDelete = event => {
-    const articleID = this.props.match.params.articleID;
-    axios.delete(`http://127.0.0.1:8000/api/${articleID}/`);
-    this.props.history.push("/");
-    this.forceUpdate();
+    if (this.props.token != null) {
+      const articleID = this.props.match.params.articleID;
+      axios.defaults.headers = {
+        "Content-Type": "application/json",
+        Authorization: this.props.token
+      };
+      axios.delete(`http://127.0.0.1:8000/api/${articleID}/`);
+      this.props.history.push("/");
+      this.forceUpdate();
+    } else {
+      //SHOW MESSAGE HERE
+    }
   };
 
   render() {
@@ -46,4 +60,10 @@ class ArticleDetail extends React.Component {
   }
 }
 
-export default ArticleDetail;
+const mapStateToProps = state => {
+  return {
+    token: state.token
+  };
+};
+
+export default connect(mapStateToProps)(ArticleDetail);
